python-backend/requirements.txt
Pillow
pillow-heif
pymupdf
pypdf
pandas
openpyxl
pyinstaller
pdf2docx
fastapi
uvicorn
python-multipart
rembg
onnxruntime
pooch
pdfplumber
pikepdf
pillow-heif
pyhanko[crypto]
requests
# Office conversions
python-docx>=1.1.0
python-pptx>=0.6.23
reportlab>=4.0.0
# HTML to PDF
weasyprint>=60.0
# OCR
pytesseract>=0.3.10
pdf2image>=1.16.3

